feat: Implement StockPipeline for stock/ETF/dividend processing with Google Sheets sync
- Added StockPipeline class to handle fetching, decrypting, analyzing, and syncing stock transactions.
- Integrated GeminiClient for PDF analysis and SheetsClient for Google Sheets synchronization.
- Created CSV parsing and writing utilities for transaction data management.
- Updated prompts for extracting transaction details from PDFs in CSV format.
- Refactored existing code into modular utility functions for better maintainability.
- Removed legacy stock_pipeline.py in favor of the new structured approach.
