+++ .gitignore
@@ -6,3 +6,4 @@ node_modules/
 attachments/*
 __pycache__/
 logs/
+.github/
