+++ brokers/__init__.py
@@ -1,0 +1,1 @@
+"""Broker/exchange adapters — Strategy pattern for data sources."""
