+++ clients/__init__.py
@@ -1,0 +1,1 @@
+"""External service client wrappers."""
