+++ main.py
@@ -1,0 +1,111 @@
+﻿"""CLI entry point for the integrated-asset pipeline.
+
+Usage:
+  python main.py                    # run all stock stages
+  python main.py --card             # run all card stages
+  python main.py --fetch            # stock fetch only
+  python main.py --since 2026/01/01 # custom date range
+  python main.py --sync             # stock sync to Google Sheets only
+"""
+
+from __future__ import annotations
+
+import argparse
+
+from dotenv import load_dotenv
+
+
+def _parse_args() -> argparse.Namespace:
+    parser = argparse.ArgumentParser(
+        description="Stock & credit-card statement processor.",
+        epilog=(
+            "Examples:\n"
+            "  python main.py                          # run all stock stages\n"
+            "  python main.py --since 2026/01/01       # fetch since a specific date\n"
+            "  python main.py --analyze                # only Gemini analysis\n"
+            "  python main.py --card                   # credit card pipeline\n"
+            "  python main.py --card --analyze         # card analyze only\n"
+            "  python main.py --sync                   # sync CSV to Google Sheet\n"
+        ),
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+    )
+    parser.add_argument(
+        "--fetch", action="store_true", help="Stage 1: Fetch attachments from Gmail"
+    )
+    parser.add_argument(
+        "--decrypt", action="store_true", help="Stage 2: Decrypt PDF attachments"
+    )
+    parser.add_argument(
+        "--analyze", action="store_true", help="Stage 3: Analyze PDFs with Gemini"
+    )
+    parser.add_argument(
+        "--card",
+        action="store_true",
+        help="Run credit-card pipeline instead of stock pipeline",
+    )
+    parser.add_argument(
+        "--since",
+        type=str,
+        metavar="YYYY/MM/DD",
+        help="Only fetch emails after this date (default: 7 days ago)",
+    )
+    parser.add_argument(
+        "--sync",
+        action="store_true",
+        help="Stage 4: Sync transactions.csv to Google Sheet",
+    )
+    parser.add_argument(
+        "--debug",
+        action="store_true",
+        help="Print raw Gemini responses and parsed rows",
+    )
+    return parser.parse_args()
+
+
+def main() -> None:
+    load_dotenv()
+    args = _parse_args()
+
+    if args.card:
+        from pipelines.card import CardPipeline
+
+        pipeline = CardPipeline()
+        run_all = not (args.fetch or args.decrypt or args.analyze)
+
+        if run_all:
+            pipeline.run_all(since=args.since, debug=args.debug)
+        else:
+            if args.fetch:
+                print("=== Card Stage 1: Fetching attachments ===")
+                pipeline.fetch(since=args.since)
+            if args.decrypt:
+                print("=== Card Stage 2: Decrypting PDFs ===")
+                pipeline.decrypt()
+            if args.analyze:
+                print("=== Card Stage 3: Analyzing with Gemini ===")
+                pipeline.analyze(debug=args.debug)
+    else:
+        from pipelines.stock import StockPipeline
+
+        pipeline = StockPipeline()
+        run_all = not (args.fetch or args.decrypt or args.analyze or args.sync)
+
+        if run_all:
+            pipeline.run_all(since=args.since, debug=args.debug)
+        else:
+            if args.fetch:
+                print("=== Stage 1: Fetching attachments ===")
+                pipeline.fetch(since=args.since)
+            if args.decrypt:
+                print("=== Stage 2: Decrypting PDFs ===")
+                pipeline.decrypt()
+            if args.analyze:
+                print("=== Stage 3: Analyzing with Gemini ===")
+                pipeline.analyze(debug=args.debug)
+            if args.sync:
+                print("=== Stage 4: Syncing to Google Sheet ===")
+                pipeline.sync()
+
+
+if __name__ == "__main__":
+    main()
