+++ utils/__init__.py
@@ -1,0 +1,1 @@
+"""Pure utility helpers — no side effects, no external service calls."""
